requirements.txt
# langchain-core
langchain
langchain_community
langchain-ollama
ollama
camel-ai
# 'camel-ai[all]'
hnswlib
nano-graphrag